+++ R/fable.intermittent-package.R
@@ -2,5 +2,11 @@
 "_PACKAGE"
 
 ## usethis namespace: start
+#' @useDynLib fable.intermittent
+#' @importFrom Rcpp sourceCpp
+#' @importFrom fabletools report
+#' @importFrom generics glance tidy
+#' @importFrom tibble tibble
+#' @importFrom stats var
 ## usethis namespace: end
 NULL