+++ R/zzz.R
@@ -1,3 +1,0 @@
-#' @useDynLib fable.intermittent
-#' @importFrom Rcpp sourceCpp
-NULL