+++ NEWS.md
@@ -1,3 +1,17 @@
+# fable.intermittent 0.1.1
+
+## New features
+
+* Added `report()` and `tidy()` methods for all model classes.
+
+* Details included in `model_sum()` for ES-based models and `STATICDISTR()`.
+
+## Bug fixes
+
+* Fixed an error in the Tweedie quantile method (`quantile.dist_tweedie()`).
+
+* `MARWAL()` forecast distributions are now truncated at zero.
+
 # fable.intermittent 0.1.0
 
 * Initial CRAN submission.